multiagent-ai-system/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
fastapi==0.116.1
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
unicorn==2.1.3
SQLAlchemy==2.0.33
email-validator==2.2.0     
python-multipart==0.0.20
tiktoken==0.9.0
openai==1.95.1
faiss-cpu==1.11.0
tenacity==9.1.2